pyproject.toml
[project]
name = "hassio-ecoflow-cloud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "homeassistant==2024.5.5",
    "jsonpath-ng>=1.7.0",
    "paho-mqtt==2.1.0",
    "protobuf==5.27.0",
]

[tool.setuptools.packages.find]
include = ["core*", "custom_components*"]
